Consolidate filter utils
JIRA: RAIL-2956
