--- libs/sdk-ui-ext/src/internal/dashboardEmbedding/mergeFiltersWithDashboard.ts
@@ -1,19 +1,0 @@
-// (C) 2020-2021 GoodData Corporation
-import { IDashboard, FilterContextItem } from "@gooddata/sdk-backend-spi";
-import { IDashboardFilter } from "./types";
-import { filterArrayToFilterContextItems } from "./utils/filters";
-
-/**
- * Gets filters merged with the filters provided by the specified dashboard.
- *
- * @param dashboard - dashboard to get the filters from
- * @param additionalFilters - filters to apply on top of the filters from the dashboard
- * @alpha
- */
-export function mergeFiltersWithDashboard(
-    dashboard: IDashboard | undefined,
-    additionalFilters: Array<IDashboardFilter | FilterContextItem>,
-): FilterContextItem[] {
-    const sanitizedAdditionalFilters = filterArrayToFilterContextItems(additionalFilters ?? []);
-    return [...(dashboard?.filterContext?.filters ?? []), ...sanitizedAdditionalFilters];
-}
